requirements.txt
numpy==2.3.2
pandas==2.3.1
Pillow==11.3.0
plotly==6.3.0
Requests==2.32.4
scipy==1.16.1
streamlit==1.48.1
Unidecode==1.4.0
